Rename for presto -> trino db
